refactor workbench performance evidence route
